Seperate Common and Customer API
Eatgo API is sepreated in 3 different module.
- common
- admin
- customer
