+++ eatgo-admin-api/src/main/java/kr/co/darcie/eatgo/EatgoAdminApiApplication.java
@@ -1,0 +1,13 @@
+package kr.co.darcie.eatgo;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+ @SpringBootApplication
+public class EatgoAdminApiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(EatgoAdminApiApplication.class, args);
+	}
+
+}
