+++ eatgo-admin-api/src/main/java/kr/co/darcie/eatgo/application/MenuItemService.java
@@ -1,0 +1,41 @@
+package kr.co.darcie.eatgo.application;
+
+import kr.co.darcie.eatgo.domain.MenuItem;
+import kr.co.darcie.eatgo.domain.MenuItemRepository;
+import kr.co.darcie.eatgo.domain.RestaurantRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class MenuItemService {
+
+    private MenuItemRepository menuItemRepository;
+
+    @Autowired
+    public MenuItemService(MenuItemRepository menuItemRepository) {
+        this.menuItemRepository = menuItemRepository;
+    }
+
+
+    public void bulkUpdate(Long restaurantId, List<MenuItem> menuItems) {
+        for (MenuItem menuItem : menuItems) {
+            update(restaurantId, menuItem);
+        }
+    }
+
+    private void update(Long restaurantId, MenuItem menuItem) {
+        if (menuItem.isDestroy()) {
+            menuItemRepository.deleteById(menuItem.getId());
+            return;
+        }
+        menuItem.setRestaurantId(restaurantId);
+        menuItemRepository.save(menuItem);
+    }
+
+    public List<MenuItem> getMenuItems(long restaurantId) {
+        List<MenuItem> menuItems = menuItemRepository.findAllByRestaurantId(restaurantId);
+        return menuItems;
+    }
+}
