+++ eatgo-admin-api/src/main/java/kr/co/darcie/eatgo/application/RestaurantService.java
@@ -1,0 +1,46 @@
+package kr.co.darcie.eatgo.application;
+
+import kr.co.darcie.eatgo.domain.*;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import javax.transaction.Transactional;
+import java.util.List;
+
+@Service
+public class RestaurantService {
+
+    private RestaurantRepository restaurantRepository;
+
+    @Autowired
+    public RestaurantService(RestaurantRepository restaurantRepository) {
+        this.restaurantRepository = restaurantRepository;
+    }
+
+    public List<Restaurant> getRestaurants() {
+        List<Restaurant> restaurants = restaurantRepository.findAll();
+
+
+        return restaurants;
+    }
+
+    public Restaurant getRestaurant(Long id){
+
+        Restaurant restaurant = restaurantRepository.findById(id)
+                .orElseThrow(() -> new RestaurantNotFoundException(id));
+
+        return restaurant;
+    };
+
+    public Restaurant addRestaurant(Restaurant restaurant) {
+        return restaurantRepository.save(restaurant);
+    }
+
+    @Transactional
+    public Restaurant updateRestaurant(Long id, String name, String address) {
+        Restaurant restaurant = restaurantRepository.findById(id).orElse(null);
+        restaurant.updateInformation(name, address);
+        return restaurant;
+    }
+
+}
