+++ eatgo-admin-api/src/main/resources/application.properties
