+++ eatgo-admin-api/src/test/java/kr/co/darcie/eatgo/EatgoAdminApiApplicationTests.java
@@ -1,0 +1,13 @@
+package kr.co.darcie.eatgo;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EatgoAdminApiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
