+++ eatgo-admin-api/src/test/java/kr/co/darcie/eatgo/interfaces/ReviewControllerTests.java
@@ -1,0 +1,45 @@
+package kr.co.darcie.eatgo.interfaces;
+
+import kr.co.darcie.eatgo.application.ReviewService;
+import kr.co.darcie.eatgo.domain.Review;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.servlet.MockMvc;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import static org.hamcrest.Matchers.containsString;
+import static org.mockito.ArgumentMatchers.contains;
+import static org.mockito.BDDMockito.given;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@RunWith(SpringRunner.class)
+@WebMvcTest(ReviewController.class)
+public class ReviewControllerTests {
+
+    @Autowired
+    MockMvc mvc;
+
+    @MockBean
+    private ReviewService reviewService;
+
+    @Test
+    public void list() throws Exception {
+        List<Review> reviews = new ArrayList<>();
+        reviews.add(Review.builder().description("Cool!").build());
+
+        given(reviewService.getReviews()).willReturn(reviews);
+
+        mvc.perform(get("/reviews"))
+            .andExpect(status().isOk())
+            .andExpect(content().string(containsString("Cool!")));
+
+    }
+}
