--- eatgo-api/src/main/java/kr/co/darcie/eatgo/EatgoApplication.java
@@ -1,13 +1,0 @@
-package kr.co.darcie.eatgo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
- @SpringBootApplication
-public class EatgoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EatgoApplication.class, args);
-	}
-
-}
