--- eatgo-api/src/main/java/kr/co/darcie/eatgo/application/ReviewService.java
@@ -1,23 +1,0 @@
-package kr.co.darcie.eatgo.application;
-
-import kr.co.darcie.eatgo.domain.Review;
-import kr.co.darcie.eatgo.domain.ReviewRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-
-@Service
-public class ReviewService {
-    private ReviewRepository reviewRepository;
-
-
-    @Autowired
-    public ReviewService(ReviewRepository reviewRepository) {
-        this.reviewRepository = reviewRepository;
-    }
-
-    public Review addReview(Long restaurantId, Review review) {
-        review.setRestaurantId(restaurantId);
-        return reviewRepository.save(review);
-    }
-}
