--- eatgo-api/src/main/java/kr/co/darcie/eatgo/domain/MenuItem.java
@@ -1,32 +1,0 @@
-package kr.co.darcie.eatgo.domain;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.*;
-import net.bytebuddy.asm.Advice;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Transient;
-
-@Entity
-@Builder
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class MenuItem {
-
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @Setter
-    private Long restaurantId;
-
-    private String name;
-
-    @Transient
-    @JsonInclude(JsonInclude.Include.NON_DEFAULT)
-    private boolean destroy;
-
-}
