--- eatgo-api/src/main/java/kr/co/darcie/eatgo/domain/ReviewRepository.java
@@ -1,11 +1,0 @@
-package kr.co.darcie.eatgo.domain;
-
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.List;
-
-public interface ReviewRepository extends CrudRepository<Review, Long> {
-    Review save(Review review);
-
-    List<Review> findAllByRestaurantId(Long restaurantId);
-}
