--- eatgo-api/src/test/java/kr/co/darcie/eatgo/EatgoApplicationTests.java
@@ -1,13 +1,0 @@
-package kr.co.darcie.eatgo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EatgoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
