--- eatgo-api/src/test/java/kr/co/darcie/eatgo/application/MenuItemServiceTests.java
@@ -1,43 +1,0 @@
-package kr.co.darcie.eatgo.application;
-
-import kr.co.darcie.eatgo.domain.MenuItem;
-import kr.co.darcie.eatgo.domain.MenuItemRepository;
-import org.junit.Before;
-import org.junit.Test;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import static org.junit.Assert.*;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-
-public class MenuItemServiceTests {
-
-    private MenuItemService menuItemService;
-
-    @Mock
-    private MenuItemRepository menuItemRepository;
-
-    @Before
-    public void setUp() {
-        MockitoAnnotations.initMocks(this);
-
-        menuItemService = new MenuItemService(menuItemRepository);
-    }
-
-    @Test
-    public void bulkUpdate() {
-        List<MenuItem> menuItems = new ArrayList<MenuItem>();
-        menuItems.add(MenuItem.builder().name("Kimchi").build());
-        menuItems.add(MenuItem.builder().id(12L).name("Gukbap").build());
-        menuItems.add(MenuItem.builder().id(1004L).destroy(true).build());
-        menuItemService.bulkUpdate(1L, menuItems);
-        verify(menuItemRepository, times(2)).save(any());
-        verify(menuItemRepository, times(1)).deleteById(eq(1004L));
-    }
-}
