+++ eatgo-common/src/main/java/kr/co/darcie/eatgo/domain/RestaurantNotFoundException.java
