+++ eatgo-common/src/main/java/kr/co/darcie/eatgo/domain/ReviewRepository.java
@@ -1,0 +1,13 @@
+package kr.co.darcie.eatgo.domain;
+
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.List;
+
+public interface ReviewRepository extends CrudRepository<Review, Long> {
+    Review save(Review review);
+
+    List<Review> findAll();
+
+    List<Review> findAllByRestaurantId(Long restaurantId);
+}
