+++ eatgo-common/src/test/java/kr/co/darcie/eatgo/domain/RestaurantTests.java
