+++ eatgo-customer-api/src/main/java/kr/co/darcie/eatgo/EatgoCustomerApiApplication.java
@@ -1,0 +1,13 @@
+package kr.co.darcie.eatgo;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EatgoCustomerApiApplication {
+
+   public static void main(String[] args) {
+       SpringApplication.run(EatgoCustomerApiApplication.class, args);
+   }
+
+}
