+++ eatgo-customer-api/src/main/java/kr/co/darcie/eatgo/interfaces/MenuItemController.java
@@ -1,0 +1,27 @@
+package kr.co.darcie.eatgo.interfaces;
+
+import kr.co.darcie.eatgo.application.MenuItemService;
+import kr.co.darcie.eatgo.domain.MenuItem;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+@RestController
+public class MenuItemController {
+
+    @Autowired
+    private MenuItemService menuItemService;
+
+    @PatchMapping("/restaurants/{restaurantId}/menuitems")
+    public String bulkUpdate(
+            @PathVariable Long restaurantId,
+            @RequestBody List<MenuItem> menuItems
+    ) {
+        menuItemService.bulkUpdate(restaurantId, menuItems);
+        return "";
+    }
+}
