+++ eatgo-customer-api/src/main/java/kr/co/darcie/eatgo/interfaces/RestaurantController.java
@@ -1,0 +1,38 @@
+package kr.co.darcie.eatgo.interfaces;
+
+import kr.co.darcie.eatgo.application.RestaurantService;
+import kr.co.darcie.eatgo.domain.Restaurant;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import javax.validation.Valid;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.util.List;
+
+@CrossOrigin
+@RestController
+public class RestaurantController {
+
+
+    @Autowired
+    private RestaurantService restaurantService;
+
+
+    @GetMapping("/restaurants")
+    public List<Restaurant> list() {
+
+        List<Restaurant> restaurants = restaurantService.getRestaurants();
+
+        return restaurants;
+    }
+
+    @GetMapping("/restaurants/{id}")
+    public Restaurant detail(@PathVariable("id") Long id) {
+        Restaurant restaurant = restaurantService.getRestaurant(id);
+
+        return restaurant;
+    }
+
+}
