+++ eatgo-customer-api/src/main/java/kr/co/darcie/eatgo/interfaces/RestaurantErrorAdvice.java
@@ -1,0 +1,19 @@
+package kr.co.darcie.eatgo.interfaces;
+
+import kr.co.darcie.eatgo.domain.RestaurantNotFoundException;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ControllerAdvice
+public class RestaurantErrorAdvice {
+
+    @ResponseBody
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    @ExceptionHandler(RestaurantNotFoundException.class)
+    public String handleNotFound() {
+        return "{}";
+    }
+}
