+++ eatgo-customer-api/src/main/java/kr/co/darcie/eatgo/interfaces/WelcomeController.java
@@ -1,0 +1,15 @@
+package kr.co.darcie.eatgo.interfaces;
+
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class WelcomeController {
+
+    @GetMapping("/")
+    public String hello() {
+        return "Hello, world!!!";
+    }
+
+}
