+++ eatgo-customer-api/src/test/java/kr/co/darcie/eatgo/EatgoCustomerApiApplicationTests.java
@@ -1,0 +1,13 @@
+package kr.co.darcie.eatgo;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EatgoCustomerApiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
